Move AfterStarting and LoadOrCreateUiConfig to App class
