+++ WalletWasabi.Fluent.Desktop/Program.cs
@@ -108,7 +108,10 @@ public class Program
 		Logger.LogWarning(e);
 
 	[SuppressMessage("CodeQuality", "IDE0051:Remove unused private members", Justification = "Required to bootstrap Avalonia's Visual Previewer")]
-	private static AppBuilder BuildAvaloniaApp() => AppBuilder.Configure(() => new App()).UseReactiveUI().SetupAppBuilder();
+	private static AppBuilder BuildAvaloniaApp()
+	{
+		return AppBuilderDesktopExtension.SetupAppBuilder(AppBuilder.Configure(() => new App()).UseReactiveUI());
+	}
 
 	/// <summary>
 	/// Sets up and initializes the crash reporting UI.
@@ -144,67 +147,6 @@ public static class WasabiAppExtensions
 {
 	public static async Task<ExitCode> RunAsGuiAsync(this WasabiApplication app)
 	{
-		return await app.RunAsync(afterStarting: () => AfterStarting(app));
-	}
-
-	public static Task AfterStarting(this WasabiApplication app)
-	{
-		RxApp.DefaultExceptionHandler = Observer.Create<Exception>(ex =>
-		{
-			if (Debugger.IsAttached)
-			{
-				Debugger.Break();
-			}
-
-			Logger.LogError(ex);
-
-			RxApp.MainThreadScheduler.Schedule(() =>
-				throw new ApplicationException("Exception has been thrown in unobserved ThrownExceptions", ex));
-		});
-
-		Logger.LogInfo("Wasabi GUI started.");
-		bool runGuiInBackground = app.AppConfig.Arguments.Any(arg => arg.Contains(StartupHelper.SilentArgument));
-		UiConfig uiConfig = LoadOrCreateUiConfig(Config.DataDir);
-		Services.Initialize(app.Global!, uiConfig, app.SingleInstanceChecker, app.TerminateService);
-
-		using CancellationTokenSource stopLoadingCts = new();
-
-		AppBuilder appBuilder = AppBuilder
-			.Configure(() => new App(
-				backendInitialiseAsync: async () =>
-				{
-					// macOS require that Avalonia is started with the UI thread. Hence this call must be delayed to this point.
-					await app.Global!.InitializeNoWalletAsync(app.TerminateService, stopLoadingCts.Token)
-						.ConfigureAwait(false);
-
-					// Make sure that wallet startup set correctly regarding RunOnSystemStartup
-					await StartupHelper.ModifyStartupSettingAsync(uiConfig.RunOnSystemStartup).ConfigureAwait(false);
-				}, startInBg: runGuiInBackground))
-			.UseReactiveUI()
-			.SetupAppBuilder()
-			.AfterSetup(_ => ThemeHelper.ApplyTheme(uiConfig.DarkModeEnabled ? Theme.Dark : Theme.Light));
-
-		if (app.TerminateService.CancellationToken.IsCancellationRequested)
-		{
-			Logger.LogDebug("Skip starting Avalonia UI as requested the application to stop.");
-			stopLoadingCts.Cancel();
-		}
-		else
-		{
-			// TODO: Refactor as on mobile we do not run Desktop lifetime.
-			appBuilder.StartWithClassicDesktopLifetime(app.AppConfig.Arguments);
-		}
-
-		return Task.CompletedTask;
-	}
-
-	private static UiConfig LoadOrCreateUiConfig(string dataDir)
-	{
-		Directory.CreateDirectory(dataDir);
-
-		UiConfig uiConfig = new(Path.Combine(dataDir, "UiConfig.json"));
-		uiConfig.LoadFile(createIfMissing: true);
-
-		return uiConfig;
+		return await app.RunAsync(afterStarting: () => App.AfterStarting(app, AppBuilderDesktopExtension.SetupAppBuilder));
 	}
 }
