+++ WalletWasabi.Fluent/App.axaml.cs
@@ -1,13 +1,20 @@
+using System.Diagnostics;
+using System.IO;
 using System.Linq;
+using System.Reactive;
 using System.Reactive.Concurrency;
 using System.Runtime.InteropServices;
+using System.Threading;
 using System.Threading.Tasks;
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Markup.Xaml;
+using Avalonia.ReactiveUI;
 using NBitcoin;
 using ReactiveUI;
+using WalletWasabi.Daemon;
+using WalletWasabi.Fluent.Helpers;
 using WalletWasabi.Fluent.Models;
 using WalletWasabi.Fluent.Models.ClientConfig;
 using WalletWasabi.Fluent.Models.FileSystem;
@@ -15,6 +22,7 @@ using WalletWasabi.Fluent.Models.UI;
 using WalletWasabi.Fluent.Models.Wallets;
 using WalletWasabi.Fluent.ViewModels;
 using WalletWasabi.Fluent.ViewModels.SearchBar.Sources;
+using WalletWasabi.Logging;
 
 namespace WalletWasabi.Fluent;
 
@@ -118,6 +126,68 @@ public class App : Application
 			this.AttachDevTools();
 		}
 #endif
+	}
+
+	public static Task AfterStarting(WasabiApplication app, Func<AppBuilder, AppBuilder> setupAppBuilder)
+	{
+		RxApp.DefaultExceptionHandler = Observer.Create<Exception>(ex =>
+		{
+			if (Debugger.IsAttached)
+			{
+				Debugger.Break();
+			}
+
+			Logger.LogError(ex);
+
+			RxApp.MainThreadScheduler.Schedule(() =>
+				throw new ApplicationException("Exception has been thrown in unobserved ThrownExceptions", ex));
+		});
+
+		Logger.LogInfo("Wasabi GUI started.");
+		bool runGuiInBackground = app.AppConfig.Arguments.Any(arg => arg.Contains(StartupHelper.SilentArgument));
+		UiConfig uiConfig = LoadOrCreateUiConfig(Config.DataDir);
+		Services.Initialize(app.Global!, uiConfig, app.SingleInstanceChecker, app.TerminateService);
+
+		using CancellationTokenSource stopLoadingCts = new();
+
+		AppBuilder appBuilder = AppBuilder
+			.Configure(() => new App(
+				backendInitialiseAsync: async () =>
+				{
+					// macOS require that Avalonia is started with the UI thread. Hence this call must be delayed to this point.
+					await app.Global!.InitializeNoWalletAsync(app.TerminateService, stopLoadingCts.Token)
+						.ConfigureAwait(false);
+
+					// Make sure that wallet startup set correctly regarding RunOnSystemStartup
+					await StartupHelper.ModifyStartupSettingAsync(uiConfig.RunOnSystemStartup).ConfigureAwait(false);
+				}, startInBg: runGuiInBackground))
+			.AfterSetup(_ => ThemeHelper.ApplyTheme(uiConfig.DarkModeEnabled ? Theme.Dark : Theme.Light))
+			.UseReactiveUI();
+
+		appBuilder = setupAppBuilder(appBuilder);
+
+		if (app.TerminateService.CancellationToken.IsCancellationRequested)
+		{
+			Logger.LogDebug("Skip starting Avalonia UI as requested the application to stop.");
+			stopLoadingCts.Cancel();
+		}
+		else
+		{
+			// TODO: Refactor as on mobile we do not run Desktop lifetime.
+			appBuilder.StartWithClassicDesktopLifetime(app.AppConfig.Arguments);
+		}
+
+		return Task.CompletedTask;
+	}
+
+	private static UiConfig LoadOrCreateUiConfig(string dataDir)
+	{
+		Directory.CreateDirectory(dataDir);
+
+		UiConfig uiConfig = new(Path.Combine(dataDir, "UiConfig.json"));
+		uiConfig.LoadFile(createIfMissing: true);
+
+		return uiConfig;
 	}
 
 	private static void OnExit(UiContext uiContext)
